Fix arena matchmaking underflow for sub-level-20 players
CharacterLevel is unsigned, so the sub-80 match-level window computed
`CharacterLevel - 20` in unsigned arithmetic. For players below level 20
this underflowed to a huge value, defeating the Math.Max(..., 1) floor and
producing an impossibly high minMatchLevel. No opponent could satisfy the
`>= minMatchLevel` filter, so a queue full of low-level players never
produced a match. Do the subtraction in signed space so the floor holds.

+++ Source/ACE.Server/Managers/ArenaManager.cs
@@ -285,7 +285,9 @@ namespace ACE.Server.Managers
                 else if (firstArenaPlayer.CharacterLevel < 80)
                 {
                     maxMatchLevel = firstArenaPlayer.CharacterLevel + 20;
-                    minMatchLevel = Math.Max(firstArenaPlayer.CharacterLevel - 20, 1);
+                    // CharacterLevel is unsigned; do the subtraction in signed space so the
+                    // Math.Max(..., 1) floor holds for players below level 20 instead of underflowing.
+                    minMatchLevel = (uint)Math.Max((int)firstArenaPlayer.CharacterLevel - 20, 1);
                 }
 
                 var otherPlayers = queuedPlayers.Values?
